+++ app/controllers/saved_letters_controller.rb
@@ -3,6 +3,10 @@ class SavedLettersController < ApplicationController
         @saved_letters = Letter.where(id: current_user.saved_letters.all.pluck(:letter_id)).order(created_at: :desc)
     end
     def create
+        unless current_user.present?
+          redirect_to user_session_path
+          return
+        end
         saved = SavedLetter.find_by(user_id: current_user.id, letter_id: params[:letter])
         if saved.present?
           saved.destroy
